Lezione 16 - BigQuery/requirements.txt
google-cloud-bigquery
db-dtypes